update-tool-call-result
